requirements.txt
fastapi
uvicorn
alpaca-py
python-dotenv